Renamed prompt.controller to prompt-controller. Updated file refs
