+++ src/pages/login-prompt/login-prompt.html
@@ -12,7 +12,7 @@
     <!-- <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';" /> -->
     <link rel="stylesheet" href="../tailwind-out.css" />
 
-    <script src="../prompt.controller.js"></script>
+    <script src="../prompt-controller.js"></script>
   </head>
   <body class="bg-gray-50 dark:bg-gray-900">
     <section class="mx-auto flex flex-col items-center justify-center px-6 py-8 md:h-screen lg:py-0">
