+++ src/pages/prompt/prompt.html
@@ -3,6 +3,8 @@
     <head>
         <meta charset="UTF-8" />
         <link href="../tailwind-out.css" rel="stylesheet" />
+
+        <script src="../prompt-controller.js"></script>
     </head>
     <body>
         <div id="container" class="mx-auto flex flex-col justify-center px-6 py-8 h-screen bg-white dark:bg-gray-900">
@@ -15,6 +17,5 @@
                 </div>
             </form>
         </div>
-        <script src="../prompt.controller.js"></script>
     </body>
 </html>
